+++ AlgorithmMonitor/src/main/java/org/example/algorithmmonitor/src/presentation/ConsoleBoard.java
@@ -1,0 +1,26 @@
+package org.example.algorithmmonitor.src.presentation;
+
+
+import lombok.RequiredArgsConstructor;
+import org.example.algorithmmonitor.src.application.service.ProblemStatusManager;
+import org.springframework.stereotype.Component;
+
+@Component  //View Component
+@RequiredArgsConstructor
+public class ConsoleBoard {
+
+    private final ProblemStatusManager problemStatusManager;
+
+    public void printAllInfo() {
+        problemStatusManager.getAllInfo();
+        System.out.println("ConsoleStatusBoard.printOverallStatus");
+//        problemStatusManager.presentOverallStatus();
+    }
+
+    public void printUnSolvedInfo() {
+        problemStatusManager.getUnsolvedInfo();
+        System.out.println("ConsoleStatusBoard.printUnsolved");
+//        List<ProblemBox> problemBoxList = problemStatusManager.presentUnsolved();
+//        problemList.stream().forEach(System.out::println);
+    }
+}
